refactor(cmd): Trying out /warp to
