+++ src/main/java/com/mcmiddleearth/warps/paper/WarpWatcher.java
@@ -1,0 +1,4 @@
+package com.mcmiddleearth.warps.paper;
+
+public class WarpWatcher {
+}
